backend/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
pydantic==1.9.0
python-dotenv
nltk
requests
python-multipart
SQLAlchemy==1.4.45
alembic
mysqlclient    
mysql-connector-python
python-jose
passlib==1.7.4
email-validator==1.3.1
bcrypt